Fix testing for int by comparing with np.integer
Test if arrays contain ints by testing wether they are subclasses of
np.integer instead of testing whether they are subclasses of np.int64.
The latter limits the program to 64-bit architectures and ints of type
int64. The class np.integer is a superclass of int32 and int64 such
that it can be used to test if an int is of any of the two types.

Note that int32 is not a subclass of int64 and Python uses the default
int type, which depends on the architecture (32 or 64 bit). Thus, we
can not compare against np.int64 as this will fail on 32-bit machines.
Also, we can not compare against np.int_ because this just represents
the default type on the system. We may get problems if the user feeds
in data that is not the default int but nevertheless an int (e.g., if
s.o. uses int64 on a 32-bit architecture).

+++ idtxl/estimators_mi.py
@@ -294,9 +294,9 @@ def jidt_discrete(self, var1, var2, opts=None):
         var1 = utils.discretise_max_ent(var1, num_discrete_bins)
         var2 = utils.discretise_max_ent(var2, num_discrete_bins)
     elif (discretise_method == 'none'):  # check if data is really discretised
-        assert issubclass(var1.dtype.type, np.int64), ('No discretisation '
+        assert issubclass(var1.dtype.type, np.integer), ('No discretisation '
                'requested, but input 1 is not an integer numpy array.')
-        assert issubclass(var2.dtype.type, np.int64), ('No discretisation '
+        assert issubclass(var2.dtype.type, np.integer), ('No discretisation '
                'requested, but input 2 is not an integer numpy array.')
         assert min(var1) >= 0, 'Minimum of input 1 is smaller than 0.'
         assert min(var2) >= 0, 'Minimum of input 1 is smaller than 0.'
